requirements.txt
asgiref==3.8.1
Django==5.0.4
django-quill-editor==0.1.40
sqlparse==0.5.0
tzdata==2024.1
